requirements.txt
fastapi
uvicorn
python-multipart
deepgram-sdk
pypdf
python-docx
python-dotenv
openai